The latest day's worth of intraday data for a given cryptocurrency
Source: R/latest.R

This returns only the most recent day of intraday data for the supplied ticker(s).

riingo_crypto_latest(ticker, resample_frequency = "1min",
  base_currency = NULL, exchanges = NULL, convert_currency = NULL,
  raw = FALSE)

Arguments

ticker
One or more cryptocurrency tickers. Specified as "btcusd" for bitcoin quoted in USD. A character vector.

resample_frequency
For Tiingo data, a character specified as one of: "daily", "monthly", "quarterly" or "yearly". For IEX data, a character specified at the "min" or "hour" frequencies in the form: "1min", "5min", or "2hour". For Crypto data, a character specified at the "min", "hour" or "day" frequencies similar to IEX.

base_currency
Instead of ticker you may pass a base currency. This selects all currencies with that base currency. For example if `base_currency="btc"`` tickers btcusd, btcjpy, btceur, etc.. will all be returned.

exchanges
If you would like to limit the query to a subset of exchanges, pass a comma-separated list of exchanges to select. Example) "POLONIEX, GDAX"

convert_currency
This parameter will convert the return data into another fx rate. For example if querying BTCUSD and convert_currency is 'cure', the bitcoin prices will be converted to CureCoin prices. Setting this to a value will add fxOpen, fxHigh, fxLow, fxClose, fxVolumeNotional, and fxRate accordingly. fxRate is the rate used to perform the currency calculation. If exchanges is specified, the conversion rate will be calculated using the exchanges passed.

raw
If TRUE, the raw underlying data from multiple exchanges will be returned, rather than the clean prices. This is the data that calculates the aggregated prices and quotes.

Examples

# The latest available day of intraday data for QQQ
riingo_crypto_latest("btcusd")

#> # A tibble: 1,082 x 10
#>    ticker baseCurrency quoteCurrency date                 open  high   low
#>    <chr>  <chr>        <chr>         <dttm>              <dbl> <dbl> <dbl>
#>  1 btcusd btc          usd           2018-04-05 00:00:00 6794. 6795. 6777.
#>  2 btcusd btc          usd           2018-04-05 00:01:00 6779. 6787. 6775.
#>  3 btcusd btc          usd           2018-04-05 00:02:00 6782. 6784. 6777.
#>  4 btcusd btc          usd           2018-04-05 00:03:00 6779. 6781. 6773.
#>  5 btcusd btc          usd           2018-04-05 00:04:00 6776. 6783. 6775.
#>  6 btcusd btc          usd           2018-04-05 00:05:00 6779. 6783. 6778.
#>  7 btcusd btc          usd           2018-04-05 00:06:00 6779. 6780. 6775.
#>  8 btcusd btc          usd           2018-04-05 00:07:00 6773. 6782. 6772.
#>  9 btcusd btc          usd           2018-04-05 00:08:00 6783. 6793. 6782.
#> 10 btcusd btc          usd           2018-04-05 00:09:00 6792. 6806. 6791.
#> # ... with 1,072 more rows, and 3 more variables: close <dbl>, volume <dbl>,
#> #   volumeNotional <dbl>