The latest day's worth of intraday data for a given ticker
Source: R/latest.R

This returns only the most recent day of intraday data for the supplied ticker(s).

riingo_iex_latest(ticker, resample_frequency = "1min")

Arguments

ticker
One or more tickers to download data for from Tiingo. Can be a stock, mutual fund, or ETF. A character vector.

resample_frequency
For Tiingo data, a character specified as one of: "daily", "monthly", "quarterly" or "yearly". For IEX data, a character specified at the "min" or "hour" frequencies in the form: "1min", "5min", or "2hour". For Crypto data, a character specified at the "min", "hour" or "day" frequencies similar to IEX.

Examples

# The latest available day of intraday data for QQQ
riingo_iex_latest("QQQ")

#> # A tibble: 429 x 6
#>    ticker date                 open  high   low close
#>    <chr>  <dttm>              <dbl> <dbl> <dbl> <dbl>
#>  1 QQQ    2018-04-11 13:34:00  161.  161.  161.  161.
#>  2 QQQ    2018-04-11 13:35:00  161.  161.  161.  161.
#>  3 QQQ    2018-04-11 13:36:00  161.  161.  161.  161.
#>  4 QQQ    2018-04-11 13:37:00  161.  161.  161.  161.
#>  5 QQQ    2018-04-11 13:38:00  161.  161.  161.  161.
#>  6 QQQ    2018-04-11 13:39:00  161.  161.  161.  161.
#>  7 QQQ    2018-04-11 13:40:00  161.  161.  161.  161.
#>  8 QQQ    2018-04-11 13:41:00  161.  161.  161.  161.
#>  9 QQQ    2018-04-11 13:42:00  161.  161.  161.  161.
#> 10 QQQ    2018-04-11 13:43:00  161.  161.  161.  161.
#> # ... with 419 more rows

riingo_iex_latest("QQQ", "1hour")

#> # A tibble: 7 x 6
#>   ticker date                 open  high   low close
#>   <chr>  <dttm>              <dbl> <dbl> <dbl> <dbl>
#> 1 QQQ    2018-04-11 14:00:00  161.  161.  160.  161.
#> 2 QQQ    2018-04-11 15:00:00  161.  162.  161.  161.
#> 3 QQQ    2018-04-11 16:00:00  161.  162.  161.  161.
#> 4 QQQ    2018-04-11 17:00:00  161.  161.  161.  161.
#> 5 QQQ    2018-04-11 18:00:00  161.  162.  161.  161.
#> 6 QQQ    2018-04-11 19:00:00  161.  161.  160.  160.
#> 7 QQQ    2018-04-11 20:00:00  160.  161.  160.  160.