Ticker information
Source: R/ticker_info.R

is_supported_ticker() can tell you if a given ticker is supported on Tiingo. supported_tickers() returns a tibble listing every available ticker.

is_supported_ticker(ticker, type = "tiingo")

supported_tickers(type = "tiingo")

Arguments

ticker
The single ticker to check for on Tiingo.

type
One of: "tiingo", "iex", or "crypto".

Examples

# NOT RUN {
# VOO is supported on both Tiingo and IEX
is_supported_ticker("VOO")
is_supported_ticker("VOO", type = "iex")

# PRHSX is a mutual fund that is supported by Tiingo but not IEX
is_supported_ticker("PRHSX")
is_supported_ticker("PRHSX", type = "iex")

# BTCUSD is available
is_supported_ticker("btcusd", type = "crypto")

# }