Get stock or ETF prices from the Tiingo API
Source: R/prices.R

The Tiingo API provides a large feed of historical data at the daily (and monthly, quarterly, or yearly) level.

riingo_prices(ticker, start_date = NULL, end_date = NULL,
  resample_frequency = "daily")

Arguments

ticker
One or more tickers to download data for from Tiingo. Can be a stock, mutual fund, or ETF. A character vector.

start_date
The first date to download data for. A character in the form YYYY-MM-DD, or a Date variable. The default is to download 1 year's worth of data.

end_date
The last date to download data for. A character in the form YYYY-MM-DD, or a Date variable.

resample_frequency
For Tiingo data, a character specified as one of: "daily", "monthly", "quarterly" or "yearly". For IEX data, a character specified at the "min" or "hour" frequencies in the form: "1min", "5min", or "2hour". For Crypto data, a character specified at the "min", "hour" or "day" frequencies similar to IEX.

Details

Multiple downloads are done sequentially, meaning that downloading 5 tickers costs 5 requests against your usage limits. Sadly Tiingo does not support batch downloads at the moment.

Tiingo supplied documentation regarding the resample frequency:

daily: Values returned as daily periods, with a holiday calendar

weekly: Values returned as weekly data, with days ending on Friday

monthly: Values returned as monthly data, with days ending on the last standard business day (Mon-Fri) of each month

annually: Values returned as annual data, with days ending on the last standard business day (Mon-Fri) of each year

Note, that if you choose a value in-between the resample period for weekly, monthly, and daily, the start date rolls back to consider the entire period. For example, if you choose to resample weekly, but your "start_date" parameter is set to Wednesday of that week, the start_date will be adjusted to Monday, so the entire week is captured. Another example is if you send a start_date mid-month, we roll back the start_date to the beginning of the month.

Similarly, if you provide an end_date, and it's midway through the period, we roll-forward the date to capture the whole period. In the above example, if the end date is set to a wednesday with a weekly resample, the end date is rolled forward to the Friday of that week.

Examples

# Downloading 1 year's worth of prices for AAPL
riingo_prices("AAPL")

#> # A tibble: 252 x 14
#>    ticker date                close  high   low  open   volume adjClose adjHigh
#>    <chr>  <dttm>              <dbl> <dbl> <dbl> <dbl>    <int>    <dbl>   <dbl>
#>  1 AAPL   2017-04-11 00:00:00  142.  143.  140.  143. 30379376     139.    141.
#>  2 AAPL   2017-04-12 00:00:00  142.  142.  141.  142. 20350000     140.    140.
#>  3 AAPL   2017-04-13 00:00:00  141.  142.  141.  142. 17822880     139.    140.
#>  4 AAPL   2017-04-17 00:00:00  142.  142.  141.  141. 16582094     140.    140.
#>  5 AAPL   2017-04-18 00:00:00  141.  142.  141.  141. 14697544     139.    140.
#>  6 AAPL   2017-04-19 00:00:00  141.  142.  140.  142. 17328375     138.    140.
#>  7 AAPL   2017-04-20 00:00:00  142.  143.  141.  141. 23319562     140.    141.
#>  8 AAPL   2017-04-21 00:00:00  142.  143.  142.  142. 17320928     140.    140.
#>  9 AAPL   2017-04-24 00:00:00  144.  144.  143.  144. 17116599     141.    142.
#> 10 AAPL   2017-04-25 00:00:00  145.  145.  144.  144. 18216472     142.    143.
#> # ... with 242 more rows, and 5 more variables: adjLow <dbl>, adjOpen <dbl>,
#> #   adjVolume <int>, divCash <dbl>, splitFactor <dbl>

# Downloading a range of data, using 2 tickers
riingo_prices(c("AAPL", "MSFT"), "1999-01-01", "2005-01-01")

#> # A tibble: 3,016 x 14
#>    ticker date                close  high   low  open   volume adjClose adjHigh
#>    <chr>  <dttm>              <dbl> <dbl> <dbl> <dbl>    <int>    <dbl>   <dbl>
#>  1 AAPL   1999-01-04 00:00:00  41.2  42.2  40.0  42.1  8507900     1.32    1.35
#>  2 AAPL   1999-01-05 00:00:00  43.3  43.9  41.5  41.9 12590300     1.38    1.40
#>  3 AAPL   1999-01-06 00:00:00  41.8  44.1  41.0  44.1 12040800     1.33    1.41
#>  4 AAPL   1999-01-07 00:00:00  45.0  45.1  42.1  42.2 12759100     1.43    1.44
#>  5 AAPL   1999-01-08 00:00:00  45.0  46.9  44.0  46.6  6061000     1.43    1.49
#>  6 AAPL   1999-01-11 00:00:00  45.9  46.1  44.9  45.8  5008700     1.46    1.47
#>  7 AAPL   1999-01-12 00:00:00  46.1  46.6  44.1  46.3  7328000     1.47    1.49
#>  8 AAPL   1999-01-13 00:00:00  46.5  47.3  42.2  42.9  9355500     1.48    1.51
#>  9 AAPL   1999-01-14 00:00:00  41.4  46.0  41.1  45.5 15391600     1.32    1.47
#> 10 AAPL   1999-01-15 00:00:00  41.3  42.1  40.0  41.8  8982200     1.32    1.34
#> # ... with 3,006 more rows, and 5 more variables: adjLow <dbl>, adjOpen <dbl>,
#> #   adjVolume <int>, divCash <dbl>, splitFactor <dbl>

# Monthly data
riingo_prices(c("AAPL", "MSFT"), "1999-01-01", "2005-01-01", "monthly")

#> # A tibble: 146 x 14
#>    ticker date                close  high   low  open   volume adjClose adjHigh
#>    <chr>  <dttm>              <dbl> <dbl> <dbl> <dbl>    <int>    <dbl>   <dbl>
#>  1 AAPL   1999-01-29 00:00:00  41.2  47.3  37.1  42.1   1.35e8     1.31    1.51
#>  2 AAPL   1999-02-26 00:00:00  34.8  41.9  34.5  41.7   7.25e7     1.11    1.34
#>  3 AAPL   1999-03-31 00:00:00  35.9  37.1  32.0  34.8   8.89e7     1.15    1.18
#>  4 AAPL   1999-04-30 00:00:00  46.0  47.1  33.5  36.1   1.42e8     1.47    1.50
#>  5 AAPL   1999-05-31 00:00:00  44.1  50.0  40.9  46.1   7.95e7     1.40    1.59
#>  6 AAPL   1999-06-30 00:00:00  46.3  49.0  42.1  45.0   6.18e7     1.48    1.56
#>  7 AAPL   1999-07-30 00:00:00  55.7  56.6  45.2  46.3   1.02e8     1.78    1.81
#>  8 AAPL   1999-08-31 00:00:00  65.2  65.9  52.1  55.6   8.25e7     2.08    2.10
#>  9 AAPL   1999-09-30 00:00:00  63.3  80.1  57.4  67.0   1.69e8     2.02    2.55
#> 10 AAPL   1999-10-29 00:00:00  80.1  81.1  59.5  62.1   1.29e8     2.55    2.58
#> # ... with 136 more rows, and 5 more variables: adjLow <dbl>, adjOpen <dbl>,
#> #   adjVolume <dbl>, divCash <dbl>, splitFactor <dbl>